pastas indesejadas removidas e tabelas unificadas
