+++ .gitignore
@@ -1,2 +1,8 @@
-node_modules
+
+node_modules/
 .env
+=======
+# Logs
+logs
+*.log
+
