--- Acessibilidade/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "Acessibilidade",
-  "version": "0.0.0",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "dependencies": {
-    "vue": "^3.0.4"
-  },
-  "devDependencies": {
-    "vite": "^1.0.0-rc.13",
-    "@vue/compiler-sfc": "^3.0.4"
-  }
-}
