--- Acessibilidade/src/pages/Index.tsx
@@ -1,5 +1,0 @@
-export default function home() {
-    return {
-        <h1 >Hello World </h1>
-    }
-}
