--- Acessibilidade/src/pages/_App.tsx
@@ -1,3 +1,0 @@
-function myApp() {
-    
-}
