--- GymManager/.gitignore
@@ -1,8 +1,0 @@
-
-node_modules/
-.env
-=======
-# Logs
-logs
-*.log
-
