--- GymManager/README.md
@@ -1,129 +1,0 @@
-# GymManager - Sistema de Gestão de Academia
-
-## Descrição
-Este projeto tem como objetivo informatizar os processos administrativos de uma academia, substituindo o controle manual por um sistema web integrado.
-
-Atualmente, a academia utiliza fichas físicas para cadastro, avaliações e controle de atividades, dificultando organização e acesso às informações.
-
-O sistema permitirá centralizar e automatizar essas operações.
-
----
-
-## Público-Alvo
-
-- **Administrador:** acesso total ao sistema
-- **Profissionais:** acesso aos alunos sob sua responsabilidade
-- **Alunos:** acesso ao próprio perfil
-
----
-
-## Objetivo Geral
-
-Desenvolver um sistema web completo para gestão administrativa, técnica e financeira da academia.
-
----
-
-## Tecnologias Utilizadas
-
-- **Frontend:** Vue.js
-- **Backend:** Node.js + Express
-- **Banco de Dados:** PostgreSQL
-
----
-
-## Funcionalidades (MVP)
-
-- Cadastro de alunos e profissionais
-- Sistema de login
-- Agendamento de atividades com limite de vagas
-- Registro de avaliação física
-- Controle de mensalidades
-- Check-in de alunos
-
----
-
-## Estrutura do Projeto
-
-### Backend (Node.js)
-
-Responsável pela API e regras de negócio do sistema.
-
-**Principais pastas:**
-
-- `services/` → lógica de negócio
-- `models/` → representação das entidades do banco
-- `routes/` → endpoints da API
-
----
-
-### Frontend (Vite + Vue.js)
-
-Responsável pela interface do usuário.
-
-**Principais pastas:**
-
-- `src/` → código principal da aplicação
-- `assets/` → imagens e estilos
-- `components/` → componentes reutilizáveis
-- `views/` → páginas do sistema
-- `services/` → comunicação com backend
-
----
-
-### 📄 Documentação
-
-- `docs/DER.jpg` → Diagrama Entidade-Relacionamento (imagem)
-- `docs/DER.brM3` → Diagrama Entidade-Relacionamento executável do brMOdelo
-
-O DER representa a estrutura do banco de dados do sistema, incluindo:
-
-- Usuários
-- Alunos
-- Profissionais
-- Treinos
-- Avaliações físicas
-- Check-ins
-- Mensalidades
-- Agendamentos
-
-Este diagrama é fundamental para entender os relacionamentos e regras do sistema.
-
----
-
-## Estratégia de Branches
-
-O projeto utiliza o padrão Git Flow:
-
-- `main` → versão estável (produção)
-- `develop` → ambiente de desenvolvimento
-- `feature/*` → novas funcionalidades
-
-Fluxo:
-1. Criar branch a partir de `develop`
-2. Desenvolver funcionalidade
-3. Abrir Pull Request
-4. Merge em `develop`
-5. Após validação → merge em `main`
-
----
-
-## Requisitos Não Funcionais
-
-- Sistema responsivo
-- Autenticação segura
-- Senhas criptografadas
-- Tempo de resposta inferior a 3 segundos
-
----
-
-## Fora do Escopo
-
-- Integração com pagamento online
-- Aplicativo mobile nativo
-- Integrações externas
-
----
-
-## Status do Projeto
-
-Em desenvolvimento (MVP)
