--- GymManager/backend/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^3.0.3",
-    "express": "^5.2.1",
-    "pg": "^8.20.0"
-  },
-  "description": "Backend do GymManager com suporte a ES Modules"
-}
