--- GymManager/frontend/src/App.vue
@@ -1,7 +1,0 @@
-<script setup>
-import HelloWorld from './components/HelloWorld.vue'
-</script>
-
-<template>
-  <HelloWorld />
-</template>
