+++ backend/package.json
@@ -2,17 +2,18 @@
   "name": "backend",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node index.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "type": "commonjs",
   "dependencies": {
     "bcryptjs": "^3.0.3",
     "express": "^5.2.1",
     "pg": "^8.20.0"
   },
-  "description": ""
+  "description": "Backend do GymManager com suporte a ES Modules"
 }
